Deduplicate analysis-script boilerplate and shared helpers (code review)
- scripts/campaignGraph.ts: shared loader (env parse, per-campaign fetch,
  minimal campaign shape, archived filter, cards/positions/edges) used by
  both ab-clustering.ts and layout-check.ts, which each duplicated ~60
  lines of it.
- Export pairKey from relations.ts and reuse it in boardLayout's pair
  collapse and the A/B script instead of three inline copies of the
  "a|b sorted" format.
- Export mulberry32 from boardLayout and use it in the A/B script so the
  harness's Louvain runs can't drift from the app's.

Verified unchanged behavior: layout-check reports identical bbox /
0 overlaps / deterministic on both campaigns; ab-clustering reproduces
the same resolution-0.9 partition.
